HACK: use standard __loongarch64_lp64 macro
For some reason, the __loongarch64__ macro defined in ffcall-abi.h did not
work for defining trampoline sizes in trampoline/trampoline.c and
callback/trampoline_r/trampoline.c.

Using __loongarch_lp64 should be fairly safe in defining a LoongArch LP64 ABI,
which was originally defined in ffcall-abi.h - for the purpose of building
this package.

The macro definition in the public header ffcall-abi.h was left intact to
ensure compatibility.